requirements.txt
aiomysql==0.2.0
bcrypt==4.2.0
cryptography==43.0.0
fastapi==0.111.1
fastapi-pagination==0.12.26
mysqlclient==2.2.4
pydantic==2.8.2
pyjwt==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
SQLAlchemy==2.0.31
sqlmodel==0.0.21
uvicorn==0.30.4
